Add files via upload
The script is used to verify the correctness of the generated yolo bounding boxes.
both the image and its correspond mask are shown side by side and the corresponding yolo_bbox was used in order to draw the bounding box on the images. It clearly shows that the bounding boxes are correctly extracted from the masks in the extract_bbox.ipynb file.